requirements.txt
python-telegram-bot==13.5
pyinstaller==4.3